Compile and use `:telemetry` stuff only if needed
