+++ lib/options.ex
@@ -98,7 +98,7 @@ defmodule Telemetria.Options do
       doc: """
       The application-specific events.
 
-      See `t::telemetry.event_prefix/0` and `t::telemetry.event_name/0`.
+      See `t:Telemetria.event_prefix/0` and `t:Telemetria.event_name/0`.
       """
     ],
     handler: [
