+++ lib/telemetria/application.ex
@@ -3,15 +3,23 @@ defmodule Telemetria.Application do
 
   use Elixir.Application
 
+  @backend Application.compile_env(:telemetria, :backend, Telemetria.Backend.Telemetry)
+
+  @doc false
+  def backend, do: @backend
+
+  @doc false
+  def telemetry?, do: Telemetria.Backend.Telemetry in List.wrap(backend())
+
+  @doc false
+  def open_telemetry?, do: Telemetria.Backend.OpenTelemetry in List.wrap(backend())
+
   @impl Elixir.Application
   def start(_type, _args) do
-    Application.ensure_all_started(:telemetry)
+    if telemetry?(), do: Application.ensure_all_started(:telemetry)
     Application.put_all_env(telemetria: Telemetria.Options.initial())
 
     opts = [strategy: :rest_for_one, name: Telemetria]
-
-    io_args =
-      if Version.match?(System.version(), ">= 1.15.0-dev"), do: {self(), "", []}, else: {"", []}
 
     throttler_args =
       :telemetria
@@ -23,19 +31,12 @@ defmodule Telemetria.Application do
       end
       |> Map.put_new(:default, {0, :none})
 
-    children = [
-      %{
-        id: Telemetria.Buffer,
-        start: {GenServer, :start_link, [StringIO, io_args, [name: Telemetria.Buffer]]}
-      },
-      {Telemetria.Throttler, throttler_args},
-      {Telemetria.Polling, Application.get_env(:telemetria, :polling, [])}
-    ]
-
+    children = if telemetry?(), do: [Telemetria.Telemetry], else: []
+    children = children ++ [{Telemetria.Throttler, throttler_args}]
     Supervisor.start_link(children, opts)
   end
 
   @impl Application
   def start_phase(:telemetry_setup, _start_type, []),
-    do: Telemetria.Instrumenter.setup()
+    do: if(telemetry?(), do: Telemetria.Instrumenter.setup(), else: :ok)
 end
